+++ database/migrations/2024_08_11_075557_create_izin_keluars_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->time('kembali', precision: 0)->nullable();
             $table->text('kegiatan')->nullable();
             $table->string('bukti')->nullable();
-            $table->bigInteger('user_id')->nullable();
+            $table->mediumInteger('user_id')->nullable();
             $table->timestamps();
         });
     }
