+++ database/migrations/2024_08_13_165759_create_kamus_anggarans_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('mak')->nullable();
             $table->string('detail')->nullable();
             $table->string('satuan', 40)->nullable();
-            $table->bigInteger('dipa_id')->nullable()->unsigned();
+            $table->mediumInteger('dipa_id')->nullable()->unsigned();
             $table->timestamps();
         });
     }
