+++ database/migrations/2024_08_13_165759_create_mata_anggarans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('mata_anggarans', function (Blueprint $table) {
             $table->id();
             $table->string('mak', 40)->nullable();
-            $table->bigInteger('dipa_id')->nullable()->unsigned();
+            $table->mediumInteger('dipa_id')->nullable()->unsigned();
             $table->timestamps();
         });
     }
