+++ database/migrations/2024_09_11_113838_create_arsip_dokumens_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('slug', 50)->nullable();
             $table->string('file')->nullable();
-            $table->bigInteger('kerangka_acuan_id')->nullable()->unsigned();
+            $table->mediumInteger('kerangka_acuan_id')->nullable()->unsigned();
             $table->timestamps();
         });
     }
