+++ database/migrations/2024_09_15_075202_create_spesifikasi_kerangka_acuans_table.php
@@ -16,10 +16,10 @@ return new class extends Migration
             $table->string('rincian', 80)->nullable();
             $table->integer('volume')->nullable()->unsigned();
             $table->string('satuan', 40)->nullable();
-            $table->bigInteger('harga_satuan')->nullable()->unsigned();
-            $table->bigInteger('total_harga')->nullable()->unsigned();
+            $table->mediumInteger('harga_satuan')->nullable()->unsigned();
+            $table->mediumInteger('total_harga')->nullable()->unsigned();
             $table->text('spesifikasi')->nullable();
-            $table->bigInteger('kerangka_acuan_id')->nullable()->unsigned();
+            $table->mediumInteger('kerangka_acuan_id')->nullable()->unsigned();
             $table->timestamps();
         });
     }
