+++ database/migrations/2024_09_22_164721_create_daftar_honor_pegawais_table.php
@@ -13,11 +13,11 @@ return new class extends Migration
     {
         Schema::create('daftar_honor_pegawais', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('volume')->nullable()->unsigned();
-            $table->bigInteger('harga_satuan')->nullable()->unsigned();
-            $table->bigInteger('persen_pajak')->nullable()->unsigned();
-            $table->bigInteger('user_id')->nullable()->unsigned();
-            $table->bigInteger('honor_kegiatan_id')->nullable()->unsigned();
+            $table->smallInteger('volume')->nullable()->unsigned();
+            $table->mediumInteger('harga_satuan')->nullable()->unsigned();
+            $table->decimal('persen_pajak', 5)->nullable()->unsigned();
+            $table->mediumInteger('user_id')->nullable()->unsigned();
+            $table->mediumInteger('honor_kegiatan_id')->nullable()->unsigned();
             $table->timestamps();
         });
     }
