+++ database/migrations/2024_10_09_092649_create_bast_mitras_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('bast_mitras', function (Blueprint $table) {
             $table->id();
             $table->date('tanggal_bast')->nullable();
-            $table->bigInteger('kontrak_mitra_id')->nullable()->unsigned();
+            $table->mediumInteger('kontrak_mitra_id')->nullable()->unsigned();
             $table->timestamps();
         });
     }
